Split files again
